Use Task-based async model instead of BeginX/EndX pattern
The Task based model enables us to use a more natural way to reason
about the asynchronous part of the program. Previously the async parts
were being done in two places, now all the async parts are in one place.

The logic is also easier to grok: Listening for packets is done by
looping around, parsing received data and exiting the loop once the
socket is disposed.